+++ eslint.config.js
@@ -26,8 +26,8 @@ export default [
           varsIgnorePattern: "^_",
         },
       ],
-      "simple-import-sort/imports": "error",
-      "simple-import-sort/exports": "error",
+      "simple-import-sort/imports": "warn",
+      "simple-import-sort/exports": "warn",
     },
     plugins: {
       "simple-import-sort": simpleImportSort,
